+++ docs/CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to the format [MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH](https://mcforge.readthedocs.io/en/1.15.x/conventions/versioning/).
 
+## [1.16.5-0.7] - 2021.02.21
+### Changed
+- [API] Streamlined items/blocks API
+- Removed Enchanted Golden Apple from most default food groups
+- Updated tooltips for non-food items
+### Fixed
+- Fixed diet values updating on cake blocks even when players did not actually eat it
+
 ## [1.16.5-0.6] - 2021.02.20
 ### Added
 - Added five-food-group support for: